fix: address round-7 multi-model review on #1274 — third sanitization gap, wider hyperlink() contract, deduplicate URL-join across CLI/TUI
- `bindOrRebind`'s `ConflictError` handler had a third, unsanitized
  `existing_datamate_name` interpolation — round 4's sanitization sweep
  fixed the other two occurrences of this exact pattern (in
  `runBrowserHandoff` and `createThenBindOrRebind`) but missed this one in
  a third function, making round 4's "every raw name-interpolation" claim
  inaccurate. Fixed, and swept the whole file again to confirm no others
  remain (all display sites go through `stripControlChars`; the three
  `recordApprovedBinding` cache-storage calls deliberately still store the
  raw value, which is correct — sanitization belongs at display time, not
  storage time).
- `hyperlink()` now also validates `url` via the existing `isSafeHttpUrl`,
  not just sanitizing `text`. It's an exported function now (tests import
  it), so its contract is wider than its two in-file callers — both of
  which only ever pass a `buildManageUrl(...)`-derived trusted URL, but an
  external caller passing something unvalidated wouldn't get the same
  protection the `text` side already has.
- Moved `buildManageUrl` (the URL-join helper, previously duplicated as
  three near-identical private copies across `link.ts`, `workspace.tsx`,
  and `workspace-sidebar.tsx`) to `browser-handoff.ts`, beside
  `resolveWorkspaceWebUrl` — a module all three files already import for
  that function, so this doesn't cross the deliberate CLI/TUI
  self-containment boundary the way `isSafeHttpUrl`'s split exists to
  avoid. Round 3's URL-join bug needed three separate edits because of
  this exact duplication; consolidating deletes the bug class instead of
  leaving three instances of it to individually stay in sync. Moved its
  tests to `browser-handoff.test.ts` alongside the function.
- Fixed a stale/circular doc comment on `openManageUrl` that described
  itself via a "same behavior as WorkspaceLinkedDialog's open action"
  cross-reference — that action now just calls `openManageUrl` directly
  (round-1 refactor), making the comment confusing rather than useful.
- Declined (per the review's own "possibly deliberate — ignore if so"):
  the differing "open in browser" wording across `WorkspaceLinkedDialog`
  and `AlreadyLinkedDialog` — one is a post-creation success card, the
  other a you're-already-linked prompt; the differing phrasing fits the
  differing context.
